requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
cytoolz==0.11.2
discord==1.7.3
discord.ext==0.8
discord.py==1.7.3
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.1
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.15.0
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.28.1
rlp==2.0.1
six==1.16.0
toolz==0.11.2
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.7.2
